Created CardDeck_sort and CardDeck_shuffle
Used bubble sort for the  sorting algorithm and differnet if statements to track when a suit is < = or greather then the successors suit . I also rearranged the cards when  the  teh current cards rank is  greater then the successor's card's rank

The shuffle was used to randomly rearrange the current node into a random position.

I left comments, I will use doxygen once the last method is done.